Many binary bootloaders were rewritten to make use of the new bootloader role functions.
Since the introduction of the new switch:

--bootloaders

you can setup it like this:

--bootloaders=syslinux,grub-efi

.

This means that syslinux is the first bootloader and grub-efi is an extra bootloader.

This commit modifies many of these bootloaders so that these roles for these binaries was enforced:

binary_grub-efi : Either first or extra bootloader
binary_grub-legacy : First bootloader
binary_grub-pc : Either first or extra bootloader
binary_syslinux : Either first or extra bootloader

If a bootloader is tried to be used in a role that it's not meant to be used then the build fails because that might lead to a non-bootable system.

Conflicts:
	scripts/build/binary_grub-efi
	scripts/build/binary_iso