Many binary bootloaders were rewritten to make use of the new bootloaders functions.
Since the introduction of the new switch:

--bootloaders

you can setup it like this:

--bootloaders=syslinux,grub-efi

.

This means that syslinux is the primary bootloader and grub-efi is the secondary bootloader.

This commit modifies many of these bootloaders so that the current roles for these binaries was enforced:

binary_grub-efi : Secondary bootloader
binary_grub-legacy : Primary bootloader
binary_grub-pc : Primary bootloader
binary_syslinux : Primary bootloader
binary_syslinux-efi : Secondary bootloader

If a bootloader is tried to be used in a role that it's not meant to be used then it's ignored. That might lead to a non-bootable system.