Initial implementation of: binary_syslinux-efi based on latest work on live-build master branch for grub-efi support.
In order to test this new script you need to specify at

lb config:

--bootloaders=syslinux,syslinux-efi

.

This work is based on this original patch:

https://bugs.debian.org/cgi-bin/bugreport.cgi?msg=103;bug=731709;att=1;filename=Add-support-for-EFI-boot-with-syslinux-efi.patch

This is currently not working in my EFI test virtual machine.
Let's hope the original author of the patch can take a look
and, maybe fix a minor bug.